More updates

+++ src/StubGenerator.php
@@ -17,10 +17,16 @@ class StubGenerator
     protected $target;
 
     /**
+     * @var bool
+     */
+    protected $toString;
+
+    /**
      * @param string $source
      * @param string $target
      */
-    public function __construct(string $source, string $target)
+
+    public function __construct(string $source, string $target = null)
     {
         $this->source = $source;
         $this->target = $target;
@@ -33,10 +39,6 @@ class StubGenerator
      */
     public function render(array $replacements)
     {
-        if (file_exists($this->target)) {
-            throw new RuntimeException('Cannot generate file. Target ' . $this->target . ' already exists.');
-        }
-
         $contents = file_get_contents($this->source);
 
         // Standard replacements
@@ -44,12 +46,29 @@ class StubGenerator
             $contents = str_replace($tag, $replacement, $contents);
         });
 
-        $path = pathinfo($this->target, PATHINFO_DIRNAME);
+        if (!$this->toString) {
+            if (file_exists($this->target)) {
+                throw new RuntimeException('Cannot generate file. Target ' . $this->target . ' already exists.');
+            }
 
-        if (! file_exists($path)) {
-            mkdir($path, 0776, true);
+            $path = pathinfo($this->target, PATHINFO_DIRNAME);
+
+            if (! file_exists($path)) {
+                mkdir($path, 0776, true);
+            }
+
+            file_put_contents($this->target, $contents);
+        } else {
+            return $contents;
         }
+    }
 
-        file_put_contents($this->target, $contents);
+    /**
+     * Set response to string
+     */
+    public function toString()
+    {
+        $this->toString = true;
+        return $this;
     }
 }
